requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiogram==3.3.0
python-dotenv==1.0.0
aiohttp==3.9.5
